pyproject.toml
[project]
name = "mcp-oceanbase"
version = "0.1.0"
description = "A Model Context Protocol (MCP) server that enables secure interaction with OceanBase databases. This server allows AI assistants to list tables, read data, and execute SQL queries through a controlled interface, making database exploration and analysis safer and more structured."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]>=1.0.0",
    "mysql-connector-python>=9.1.0",
]
[[project.authors]]
name = "OceanBase"
email = ""

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
include = [
    "src/oceanbase_mcp_server/**/*.py",
]

[tool.hatch.build.targets.wheel]
packages = [
    "src"
]

[project.scripts]
oceanbase_mcp_server = "src.oceanbase_mcp_server:main"
